+++ src/components/verifier/UploadPanel.tsx
@@ -17,7 +17,8 @@ import {
   AccordionButton,
   useDisclosure,
 } from "@chakra-ui/react";
-import axios from "axios";
+
+import axios, { AxiosResponse } from "axios";
 
 import { useAccount } from "wagmi";
 
@@ -38,344 +39,62 @@ export interface JsonFileContentType {
   selectedRows: selectedRows[];
 }
 
+
 export function UploadPanel() {
-
-  const handleVerified = () =>{
+  const handleVerified = () => {
     toast({
-      title: 'Verified',
+      title: "Verified",
       description: "Your proof is verified!",
-      status: 'success',
+      status: "success",
       duration: 9000,
       isClosable: true,
-    })
-  }
+    });
+  };
 
-  const handleNotVerified = () =>{
+  const handleNotVerified = () => {
     toast({
-      title: 'Not Verified',
+      title: "Not Verified",
       description: "Sorry, your proof is un-verified",
-      status: 'error',
+      status: "error",
       duration: 9000,
       isClosable: true,
-    })
-
-  }
-
+    });
+  };
 
   const handleVerifyButton = async () => {
     setIsLoading(true);
     try {
-      const res = await axios.post(
-        "https://medi0backendrusty.spicybuilds.xyz/verify-proof",
+      if (jsonFileContent !== undefined) {
+        const flattenedResult = jsonFileContent.flatMap((content1) => {
+          return content1.selectedRows.map((row) => {
+            return {
+              row_title: row.selectedKey,
+              row_content: row.selectedValue,
+              commitment: content1.certHash,
+              proof: row.proof,
+            };
+          });
+        });
+        const bulkRequest = flattenedResult.map((file) =>
+          axios.post<{ valid: boolean }>(
+            "https://medi0backendrusty.spicybuilds.xyz/verify-proof",
+            file
+          )
+        );
 
-        {
-          row_title: "else",
-          row_content: "down",
-          commitment:
-            "0x3baa9ebc22be9520e2bdbba6f600398801ed7f7c9415d97d5db504d4815e32db",
-          proof: [
-            51, 21, 10, 161, 6, 150, 68, 31, 211, 154, 159, 62, 95, 145, 226,
-            152, 41, 137, 28, 175, 47, 14, 20, 73, 181, 148, 209, 77, 105, 240,
-            61, 144, 168, 120, 26, 180, 0, 254, 13, 42, 18, 193, 141, 56, 83,
-            202, 129, 152, 12, 109, 84, 195, 79, 20, 108, 240, 79, 93, 181, 252,
-            10, 239, 124, 187, 179, 58, 65, 238, 45, 78, 232, 19, 0, 3, 198,
-            200, 75, 216, 43, 85, 224, 192, 155, 110, 178, 97, 134, 26, 41, 185,
-            181, 189, 46, 13, 229, 53, 243, 55, 136, 249, 85, 173, 52, 240, 209,
-            75, 176, 49, 42, 175, 159, 199, 234, 206, 223, 93, 72, 217, 156,
-            131, 197, 34, 243, 59, 93, 204, 53, 135, 63, 116, 27, 73, 40, 72,
-            137, 122, 188, 51, 226, 224, 125, 250, 246, 241, 202, 193, 132, 96,
-            196, 161, 171, 39, 222, 165, 148, 50, 132, 137, 170, 146, 21, 181,
-            165, 235, 75, 114, 165, 126, 114, 205, 205, 239, 194, 83, 231, 231,
-            152, 108, 74, 103, 105, 175, 170, 120, 70, 39, 171, 166, 142, 251,
-            161, 13, 114, 114, 202, 128, 80, 42, 59, 229, 111, 162, 63, 113,
-            191, 136, 190, 184, 55, 219, 255, 175, 81, 33, 54, 206, 227, 218, 2,
-            77, 209, 124, 95, 7, 103, 124, 246, 163, 16, 0, 230, 12, 247, 184,
-            91, 23, 204, 82, 217, 39, 10, 182, 55, 3, 63, 114, 135, 211, 31,
-            102, 255, 76, 191, 134, 191, 53, 33, 107, 92, 165, 190, 241, 222,
-            44, 28, 13, 197, 39, 84, 21, 166, 143, 15, 75, 170, 61, 139, 112,
-            154, 189, 252, 64, 253, 176, 107, 56, 221, 43, 137, 75, 3, 213, 97,
-            152, 77, 166, 135, 235, 42, 84, 145, 15, 221, 116, 16, 221, 122,
-            206, 183, 6, 23, 26, 162, 214, 48, 127, 92, 217, 2, 51, 16, 92, 169,
-            88, 229, 114, 112, 201, 191, 155, 17, 209, 216, 25, 61, 187, 154,
-            60, 241, 49, 137, 84, 184, 181, 241, 124, 46, 227, 85, 11, 137, 134,
-            91, 17, 41, 82, 162, 71, 163, 53, 158, 33, 56, 223, 244, 140, 140,
-            130, 174, 138, 166, 208, 154, 98, 39, 11, 108, 157, 13, 133, 96,
-            119, 238, 151, 196, 167, 194, 229, 4, 9, 148, 4, 1, 194, 122, 35,
-            139, 2, 1, 65, 231, 48, 231, 58, 99, 59, 0, 10, 8, 255, 66, 222, 78,
-            34, 66, 170, 125, 13, 141, 224, 113, 35, 4, 229, 184, 252, 155, 119,
-            74, 105, 249, 16, 235, 203, 193, 43, 36, 43, 124, 211, 6, 70, 199,
-            197, 34, 238, 200, 44, 238, 156, 52, 142, 60, 113, 210, 25, 183,
-            137, 63, 126, 152, 188, 72, 166, 178, 60, 39, 90, 23, 119, 76, 223,
-            205, 168, 202, 102, 88, 128, 85, 136, 178, 80, 187, 54, 11, 226,
-            225, 143, 44, 2, 242, 96, 103, 219, 16, 20, 157, 35, 69, 3, 92, 50,
-            175, 248, 76, 245, 236, 24, 218, 142, 237, 147, 130, 44, 36, 86, 18,
-            67, 18, 202, 217, 120, 61, 94, 232, 238, 100, 6, 204, 120, 23, 143,
-            125, 2, 5, 1, 126, 77, 223, 120, 247, 85, 90, 157, 178, 103, 57, 0,
-            91, 254, 248, 189, 18, 25, 36, 121, 170, 135, 60, 31, 163, 1, 97,
-            64, 236, 195, 163, 42, 97, 25, 96, 220, 121, 169, 135, 1, 58, 93,
-            236, 94, 62, 70, 165, 107, 202, 143, 217, 16, 41, 104, 235, 21, 162,
-            60, 13, 66, 88, 1, 153, 237, 133, 80, 255, 81, 252, 80, 88, 32, 95,
-            95, 163, 57, 99, 214, 63, 116, 115, 250, 177, 230, 93, 19, 224, 75,
-            220, 9, 190, 181, 155, 199, 125, 101, 127, 54, 193, 237, 36, 150,
-            47, 176, 249, 247, 27, 204, 54, 74, 131, 27, 71, 123, 195, 54, 41,
-            127, 104, 246, 184, 47, 6, 219, 125, 122, 31, 129, 237, 194, 136,
-            212, 212, 40, 55, 28, 87, 75, 121, 6, 153, 154, 143, 121, 32, 95,
-            145, 219, 151, 207, 146, 141, 61, 162, 23, 131, 240, 184, 58, 116,
-            218, 208, 103, 106, 106, 8, 113, 171, 182, 153, 38, 26, 99, 224, 10,
-            86, 28, 65, 79, 87, 89, 11, 89, 39, 212, 218, 199, 57, 173, 99, 113,
-            153, 141, 104, 9, 58, 147, 193, 191, 43, 90, 23, 113, 151, 168, 138,
-            18, 110, 169, 186, 150, 249, 251, 46, 110, 117, 34, 50, 42, 216, 91,
-            51, 27, 2, 177, 213, 151, 39, 227, 205, 77, 153, 225, 13, 70, 16,
-            53, 92, 239, 67, 117, 46, 233, 109, 9, 202, 16, 194, 44, 99, 231,
-            30, 242, 225, 196, 230, 188, 37, 167, 105, 1, 160, 99, 9, 169, 50,
-            196, 44, 159, 174, 58, 81, 107, 240, 88, 101, 87, 7, 134, 118, 122,
-            40, 149, 68, 255, 5, 175, 44, 158, 229, 114, 201, 29, 97, 119, 37,
-            159, 98, 38, 97, 242, 27, 247, 87, 116, 150, 136, 120, 27, 62, 161,
-            49, 12, 152, 97, 61, 42, 100, 121, 16, 8, 161, 61, 70, 97, 89, 52,
-            57, 18, 48, 252, 15, 32, 177, 182, 85, 159, 161, 28, 231, 9, 71,
-            168, 189, 102, 21, 253, 79, 50, 176, 18, 10, 89, 40, 82, 147, 57,
-            49, 222, 52, 37, 110, 75, 214, 202, 37, 128, 107, 60, 73, 112, 186,
-            186, 55, 225, 255, 16, 195, 130, 211, 193, 243, 88, 47, 5, 0, 36,
-            18, 66, 183, 46, 155, 0, 247, 70, 243, 7, 65, 2, 148, 78, 130, 46,
-            109, 255, 245, 194, 117, 52, 118, 111, 33, 227, 9, 44, 3, 223, 133,
-            203, 127, 206, 168, 175, 156, 0, 175, 198, 203, 95, 253, 251, 100,
-            56, 168, 56, 211, 246, 197, 116, 225, 132, 223, 205, 6, 153, 157,
-            69, 230, 22, 51, 123, 212, 32, 12, 193, 6, 0, 175, 198, 203, 95,
-            253, 251, 100, 56, 168, 56, 211, 246, 197, 116, 225, 132, 223, 205,
-            6, 153, 157, 69, 230, 22, 51, 123, 212, 32, 12, 193, 6, 117, 126,
-            172, 24, 238, 116, 63, 167, 161, 233, 79, 185, 104, 196, 194, 207,
-            22, 108, 194, 188, 255, 26, 137, 18, 195, 11, 223, 126, 77, 14, 184,
-            59, 230, 145, 182, 140, 247, 34, 71, 113, 139, 123, 97, 138, 232,
-            71, 175, 67, 107, 178, 35, 97, 187, 195, 209, 112, 44, 54, 57, 105,
-            102, 158, 56, 41, 194, 171, 42, 181, 76, 192, 89, 255, 80, 218, 92,
-            31, 150, 98, 217, 251, 242, 216, 189, 241, 209, 2, 15, 76, 193, 105,
-            133, 163, 149, 38, 31, 45, 191, 239, 32, 110, 157, 112, 70, 246,
-            192, 122, 16, 86, 146, 24, 217, 222, 36, 95, 244, 226, 123, 129,
-            227, 16, 11, 116, 152, 254, 202, 117, 158, 7, 68, 231, 221, 94, 11,
-            90, 33, 29, 215, 96, 114, 13, 43, 32, 237, 1, 222, 137, 77, 69, 16,
-            239, 14, 48, 183, 71, 255, 72, 146, 178, 249, 60, 77, 139, 250, 117,
-            239, 182, 69, 204, 77, 122, 89, 169, 22, 97, 200, 80, 211, 207, 180,
-            36, 217, 99, 101, 104, 89, 184, 191, 23, 40, 213, 2, 54, 109, 40,
-            252, 242, 129, 42, 28, 136, 65, 36, 151, 226, 130, 230, 23, 116, 25,
-            36, 136, 163, 120, 199, 0, 21, 74, 64, 102, 185, 212, 91, 3, 30, 28,
-            2, 216, 8, 31, 47, 128, 2, 217, 165, 151, 156, 162, 199, 137, 17,
-            79, 232, 222, 101, 108, 120, 23, 51, 32, 18, 196, 111, 33, 24, 113,
-            26, 213, 12, 56, 19, 194, 248, 46, 166, 165, 50, 190, 121, 119, 247,
-            198, 78, 203, 24, 191, 220, 49, 150, 48, 25, 59, 176, 217, 124, 103,
-            185, 60, 44, 15, 26, 80, 37, 20, 117, 216, 109, 35, 200, 160, 135,
-            120, 151, 88, 33, 227, 173, 138, 54, 189, 243, 12, 115, 207, 204,
-            223, 20, 244, 176, 197, 2, 212, 150, 167, 199, 116, 79, 179, 160,
-            222, 231, 103, 162, 101, 36, 119, 75, 104, 137, 37, 158, 128, 255,
-            127, 153, 153, 49, 238, 248, 30, 41, 227, 27, 143, 41, 81, 158, 172,
-            186, 167, 116, 146, 163, 109, 13, 159, 88, 255, 153, 103, 65, 7,
-            229, 45, 148, 243, 25, 113, 202, 60, 57, 245, 252, 11, 51, 177, 24,
-            244, 45, 172, 45, 211, 38, 120, 154, 219, 118, 124, 164, 251, 248,
-            222, 58, 146, 134, 236, 111, 55, 108, 238, 150, 254, 91, 224, 66,
-            66, 42, 0, 191, 166, 237, 94, 162, 42, 225, 197, 183, 63, 151, 54,
-            238, 160, 43, 53, 48, 236, 5, 28, 197, 167, 214, 20, 200, 116, 240,
-            11, 234, 48, 49, 106, 162, 87, 9, 33, 235, 83, 165, 32, 70, 23, 136,
-            53, 96, 133, 168, 26, 60, 47, 33, 77, 255, 155, 58, 19, 234, 58,
-            153, 94, 208, 122, 14, 103, 101, 126, 236, 15, 62, 93, 48, 141, 170,
-            105, 110, 105, 43, 50, 5, 207, 202, 180, 187, 40, 64, 242, 154, 132,
-            226, 172, 93, 8, 194, 101, 51, 7, 119, 20, 238, 32, 227, 109, 228,
-            67, 116, 121, 155, 2, 94, 92, 149, 175, 64, 129, 170, 51, 49, 54,
-            213, 25, 255, 192, 7, 86, 17, 112, 32, 153, 239, 56, 101, 221, 166,
-            252, 73, 157, 125, 207, 82, 59, 27, 74, 233, 57, 124, 145, 61, 225,
-            237, 9, 136, 155, 65, 173, 141, 4, 156, 31, 57, 196, 83, 58, 81,
-            185, 67, 126, 13, 92, 35, 152, 228, 72, 149, 35, 173, 16, 201, 243,
-            143, 186, 169, 163, 233, 171, 220, 201, 225, 255, 152, 35, 56, 185,
-            67, 137, 68, 114, 174, 52, 116, 31, 53, 56, 231, 109, 67, 246, 205,
-            189, 108, 95, 101, 162, 125, 15, 244, 63, 118, 240, 175, 106, 104,
-            62, 30, 219, 35, 162, 195, 120, 86, 176, 162, 24, 33, 252, 38, 114,
-            32, 158, 72, 28, 197, 24, 39, 144, 224, 40, 159, 12, 246, 178, 111,
-            223, 211, 83, 26, 224, 184, 249, 160, 43, 125, 116, 14, 118, 38, 48,
-            131, 66, 66, 215, 71, 233, 227, 228, 66, 157, 40, 92, 176, 118, 13,
-            60, 220, 89, 5, 251, 6, 85, 19, 253, 204, 37, 124, 0, 89, 214, 140,
-            40, 159, 0, 56, 35, 75, 216, 4, 6, 59, 141, 99, 12, 183, 83, 163,
-            138, 123, 36, 67, 44, 42, 38, 97, 206, 161, 163, 202, 16, 85, 249,
-            1, 165, 183, 245, 9, 252, 192, 112, 241, 205, 180, 108, 41, 79, 147,
-            161, 236, 87, 70, 232, 51, 79, 6, 214, 26, 86, 250, 155, 71, 218,
-            54, 147, 245, 169, 196, 58, 60, 171, 155, 80, 67, 106, 230, 25, 113,
-            157, 240, 149, 161, 116, 227, 81, 64, 26, 1, 10, 249, 189, 156, 240,
-            128, 139, 20, 216, 125, 97, 240, 76, 66, 1, 180, 158, 242, 220, 143,
-            251, 107, 101, 85, 86, 105, 7, 167, 39, 238, 225, 8, 162, 125, 187,
-            202, 227, 50, 23, 252, 229, 29, 79, 171, 189, 136, 254, 31, 208,
-            176, 79, 37, 40, 11, 239, 255, 10, 72, 129, 144, 151, 252, 200, 15,
-            246, 4, 41, 23, 51, 151, 133, 121, 135, 81, 61, 52, 102, 108, 46,
-            31, 56, 173, 215, 196, 133, 112, 195, 200, 88, 198, 49, 183, 209,
-            120, 23, 63, 100, 9, 73, 7, 175, 107, 121, 67, 205, 43, 112, 65,
-            171, 251, 80, 219, 33, 69, 80, 223, 140, 224, 188, 206, 38, 6, 249,
-            27, 153, 63, 23, 62, 224, 106, 7, 63, 143, 105, 88, 48, 26, 85, 127,
-            168, 242, 183, 68, 125, 185, 5, 195, 118, 130, 105, 244, 139, 14,
-            198, 60, 218, 186, 37, 62, 15, 238, 245, 166, 158, 160, 3, 170, 24,
-            213, 77, 6, 20, 91, 2, 193, 111, 183, 63, 144, 35, 168, 36, 176,
-            100, 173, 212, 67, 218, 187, 141, 90, 23, 229, 164, 114, 164, 219,
-            170, 146, 253, 172, 196, 52, 182, 6, 138, 239, 253, 193, 11, 59,
-            242, 198, 149, 234, 90, 64, 222, 179, 231, 137, 177, 205, 18, 110,
-            105, 255, 55, 141, 232, 30, 175, 45, 65, 202, 135, 238, 13, 76, 86,
-            146, 248, 245, 219, 254, 31, 174, 28, 77, 22, 57, 111, 45, 234, 45,
-            18, 169, 245, 75, 48, 200, 125, 24, 179, 180, 101, 207, 188, 66,
-            112, 3, 95, 197, 24, 166, 61, 196, 71, 57, 145, 170, 141, 10, 5,
-            216, 202, 31, 21, 221, 163, 192, 72, 127, 27, 204, 67, 32, 220, 99,
-            203, 139, 204, 121, 233, 159, 235, 255, 34, 26, 142, 26, 196, 69,
-            203, 129, 211, 106, 135, 35, 1, 39, 4, 60, 247, 94, 105, 55, 238,
-            54, 163, 17, 137, 114, 204, 87, 156, 10, 199, 108, 61, 72, 212, 4,
-            74, 95, 115, 183, 63, 90, 221, 71, 24, 89, 175, 244, 63, 104, 230,
-            235, 197, 51, 113, 43, 122, 52, 219, 208, 188, 201, 151, 184, 194,
-            184, 126, 18, 196, 253, 24, 157, 183, 57, 63, 175, 26, 95, 11, 1,
-            107, 142, 222, 244, 195, 32, 212, 140, 14, 250, 73, 163, 123, 7, 71,
-            174, 212, 48, 235, 122, 30, 81, 248, 20, 84, 233, 29, 139, 17, 2,
-            197, 71, 145, 123, 180, 48, 33, 56, 12, 58, 117, 120, 170, 21, 198,
-            155, 94, 51, 68, 220, 59, 56, 248, 32, 182, 206, 147, 79, 248, 125,
-            38, 151, 158, 24, 107, 140, 230, 154, 39, 212, 25, 24, 65, 66, 3,
-            160, 79, 13, 110, 225, 86, 238, 104, 80, 253, 88, 50, 120, 4, 184,
-            212, 230, 15, 85, 249, 165, 141, 17, 233, 43, 210, 255, 140, 93, 89,
-            138, 81, 142, 156, 235, 11, 153, 81, 192, 91, 223, 230, 237, 178,
-            225, 212, 133, 46, 178, 8, 243, 81, 117, 42, 159, 103, 55, 109, 235,
-            83, 3, 225, 132, 106, 80, 86, 88, 254, 26, 92, 97, 167, 160, 0, 42,
-            120, 250, 41, 150, 207, 40, 11, 122, 4, 218, 254, 210, 102, 45, 121,
-            250, 95, 95, 107, 157, 110, 78, 4, 116, 246, 243, 205, 180, 199,
-            119, 244, 78, 86, 127, 176, 72, 187, 20, 37, 82, 36, 72, 163, 12,
-            62, 95, 21, 164, 231, 200, 34, 159, 245, 176, 131, 65, 141, 94, 173,
-            158, 118, 44, 65, 233, 222, 116, 98, 229, 255, 19, 62, 78, 196, 119,
-            184, 36, 215, 87, 183, 115, 204, 183, 219, 207, 190, 95, 127, 227,
-            37, 180, 69, 148, 232, 237, 197, 226, 199, 15, 6, 109, 228, 137, 63,
-            192, 92, 172, 69, 6, 11, 81, 92, 220, 237, 64, 216, 66, 158, 186,
-            12, 90, 164, 179, 192, 192, 149, 190, 178, 33, 67, 91, 254, 111, 35,
-            216, 33, 253, 91, 79, 53, 60, 123, 255, 55, 159, 43, 45, 186, 133,
-            26, 126, 19, 132, 151, 109, 63, 89, 251, 109, 124, 252, 26, 184,
-            162, 172, 141, 134, 35, 236, 4, 190, 143, 117, 219, 32, 151, 226, 1,
-            148, 171, 207, 179, 249, 240, 62, 182, 34, 85, 171, 247, 245, 57,
-            187, 181, 127, 59, 81, 229, 151, 16, 87, 218, 198, 30, 177, 216, 58,
-            53, 5, 63, 30, 111, 132, 176, 227, 189, 109, 32, 207, 66, 227, 73,
-            216, 63, 108, 116, 141, 21, 123, 71, 4, 27, 204, 198, 74, 170, 196,
-            173, 234, 234, 167, 11, 63, 244, 8, 151, 247, 127, 160, 44, 75, 248,
-            74, 174, 58, 157, 65, 37, 108, 67, 209, 41, 138, 54, 45, 91, 175,
-            104, 146, 118, 70, 17, 159, 229, 208, 186, 39, 208, 246, 216, 175,
-            12, 223, 70, 171, 155, 5, 73, 255, 177, 123, 124, 187, 252, 31, 26,
-            208, 76, 52, 41, 57, 36, 243, 36, 27, 90, 25, 233, 143, 172, 87,
-            168, 250, 200, 51, 74, 214, 224, 6, 92, 117, 162, 159, 226, 126,
-            183, 83, 59, 170, 22, 120, 32, 208, 145, 245, 244, 95, 3, 146, 191,
-            197, 107, 3, 71, 198, 41, 245, 45, 95, 224, 54, 201, 91, 222, 121,
-            39, 192, 241, 175, 24, 198, 20, 140, 111, 27, 208, 158, 128, 176,
-            198, 246, 115, 105, 15, 138, 6, 225, 73, 198, 118, 88, 126, 83, 224,
-            104, 160, 41, 20, 40, 72, 139, 28, 220, 12, 116, 83, 163, 144, 231,
-            76, 45, 57, 14, 251, 217, 132, 185, 10, 114, 161, 248, 255, 92, 159,
-            212, 104, 87, 57, 34, 24, 83, 224, 249, 26, 82, 231, 228, 123, 50,
-            211, 99, 71, 42, 229, 108, 39, 176, 106, 105, 97, 196, 90, 18, 89,
-            100, 34, 175, 188, 200, 75, 204, 140, 78, 84, 247, 1, 191, 76, 114,
-            70, 40, 5, 214, 178, 62, 184, 35, 202, 110, 156, 11, 244, 139, 140,
-            59, 230, 55, 53, 168, 175, 24, 194, 124, 39, 111, 232, 140, 48, 24,
-            148, 52, 194, 146, 196, 243, 108, 139, 182, 55, 34, 59, 114, 207,
-            87, 108, 248, 108, 89, 79, 157, 165, 36, 171, 86, 79, 169, 185, 58,
-            117, 5, 84, 75, 158, 189, 136, 200, 88, 208, 127, 153, 161, 207,
-            245, 54, 184, 176, 87, 117, 188, 222, 227, 35, 70, 18, 87, 9, 120,
-            59, 36, 88, 199, 7, 156, 16, 147, 119, 160, 88, 88, 169, 5, 162, 87,
-            200, 129, 232, 153, 238, 214, 92, 53, 138, 164, 212, 96, 169, 234,
-            21, 247, 115, 114, 64, 171, 15, 17, 148, 140, 233, 172, 153, 121,
-            245, 85, 60, 212, 24, 86, 236, 111, 7, 191, 92, 130, 160, 145, 120,
-            54, 185, 16, 47, 208, 35, 42, 179, 203, 38, 194, 202, 164, 144, 204,
-            47, 17, 153, 28, 213, 84, 14, 121, 198, 225, 147, 179, 221, 198, 9,
-            123, 0, 210, 89, 205, 80, 141, 122, 240, 159, 19, 55, 220, 20, 71,
-            172, 59, 8, 107, 22, 121, 144, 179, 249, 167, 19, 234, 45, 119, 4,
-            45, 97, 96, 244, 58, 130, 92, 205, 215, 108, 48, 11, 170, 30, 128,
-            153, 36, 124, 54, 75, 111, 179, 200, 71, 25, 11, 166, 125, 91, 193,
-            153, 1, 103, 216, 18, 55, 150, 115, 195, 153, 223, 48, 180, 99, 39,
-            45, 254, 67, 247, 207, 234, 59, 252, 87, 255, 31, 88, 78, 78, 225,
-            131, 12, 209, 110, 156, 182, 88, 88, 184, 97, 169, 129, 152, 54,
-            133, 134, 68, 21, 3, 96, 56, 65, 228, 153, 249, 164, 231, 198, 69,
-            239, 64, 17, 98, 22, 89, 43, 31, 229, 129, 211, 207, 202, 63, 51,
-            108, 89, 183, 53, 147, 3, 85, 172, 111, 67, 139, 5, 31, 0, 41, 189,
-            127, 172, 249, 207, 107, 93, 137, 155, 154, 82, 58, 74, 22, 100,
-            186, 115, 67, 21, 96, 242, 42, 60, 151, 215, 204, 6, 179, 93, 97,
-            162, 195, 138, 93, 57, 175, 165, 86, 90, 176, 63, 93, 252, 151, 162,
-            147, 239, 54, 237, 186, 198, 182, 216, 248, 24, 234, 29, 80, 236,
-            244, 53, 245, 121, 131, 241, 232, 215, 159, 250, 133, 133, 195, 189,
-            238, 80, 137, 54, 233, 188, 26, 251, 75, 128, 120, 38, 4, 58, 163,
-            235, 76, 77, 215, 128, 25, 200, 102, 106, 190, 220, 13, 202, 23, 16,
-            31, 22, 75, 187, 25, 190, 186, 166, 129, 231, 218, 25, 38, 224, 29,
-            148, 72, 247, 157, 14, 248, 248, 187, 224, 19, 20, 0, 62, 115, 185,
-            221, 204, 145, 242, 230, 17, 246, 197, 158, 239, 163, 128, 40, 174,
-            52, 100, 224, 25, 226, 80, 2, 203, 24, 75, 96, 240, 139, 253, 156,
-            116, 24, 151, 201, 156, 245, 188, 202, 248, 237, 12, 88, 69, 188,
-            251, 211, 202, 33, 95, 160, 35, 201, 255, 227, 192, 205, 55, 1, 162,
-            13, 177, 169, 170, 188, 15, 113, 128, 162, 221, 137, 18, 70, 24, 91,
-            230, 111, 118, 73, 159, 183, 217, 116, 31, 58, 134, 13, 227, 147,
-            167, 253, 47, 174, 78, 230, 88, 2, 239, 0, 40, 80, 235, 178, 151,
-            179, 78, 99, 219, 30, 62, 176, 29, 252, 68, 94, 2, 5, 45, 22, 201,
-            38, 255, 222, 118, 84, 161, 64, 40, 91, 241, 70, 121, 59, 167, 155,
-            28, 239, 211, 218, 168, 138, 1, 37, 54, 244, 48, 135, 185, 28, 19,
-            40, 163, 186, 36, 138, 17, 150, 124, 168, 14, 67, 190, 98, 232, 191,
-            250, 41, 232, 202, 68, 225, 199, 212, 123, 177, 178, 231, 203, 222,
-            132, 195, 68, 81, 152, 175, 54, 116, 42, 33, 2, 151, 239, 58, 199,
-            171, 57, 203, 189, 199, 58, 9, 116, 125, 193, 34, 211, 140, 217, 99,
-            140, 31, 52, 161, 174, 116, 234, 200, 118, 118, 69, 18, 232, 39, 16,
-            245, 247, 100, 120, 134, 93, 198, 148, 144, 113, 83, 226, 102, 189,
-            16, 117, 158, 31, 129, 49, 246, 170, 48, 206, 120, 217, 92, 71, 62,
-            16, 249, 166, 155, 67, 180, 201, 60, 114, 203, 105, 16, 8, 130, 234,
-            15, 81, 20, 101, 95, 157, 156, 151, 226, 17, 143, 51, 197, 207, 219,
-            42, 72, 14, 112, 95, 196, 103, 116, 172, 229, 70, 87, 38, 200, 246,
-            232, 254, 61, 185, 117, 19, 80, 13, 183, 183, 90, 138, 84, 119, 62,
-            251, 149, 243, 232, 109, 71, 184, 134, 37, 154, 203, 209, 249, 137,
-            63, 229, 245, 5, 219, 41, 38, 27, 180, 254, 147, 195, 164, 60, 19,
-            149, 13, 4, 71, 56, 84, 37, 78, 250, 190, 239, 80, 126, 236, 0, 65,
-            166, 4, 200, 122, 7, 210, 221, 66, 47, 62, 199, 65, 22, 147, 142,
-            42, 218, 141, 1, 242, 5, 205, 97, 125, 39, 25, 174, 179, 2, 179, 2,
-            169, 225, 132, 89, 242, 84, 90, 158, 211, 65, 25, 48, 168, 19, 166,
-            199, 242, 228, 180, 11, 124, 233, 243, 83, 173, 223, 181, 203, 166,
-            139, 101, 112, 212, 180, 242, 15, 183, 101, 212, 62, 131, 239, 184,
-            249, 143, 8, 63, 6, 46, 14, 140, 25, 111, 207, 198, 179, 79, 76,
-            217, 232, 71, 101, 173, 88, 155, 91, 45, 132, 10, 36, 154, 45, 247,
-            23, 119, 249, 160, 186, 16, 146, 63, 97, 119, 53, 129, 32, 78, 130,
-            126, 18, 99, 93, 74, 49, 182, 194, 93, 109, 79, 13, 198, 80, 188, 7,
-            192, 42, 238, 234, 163, 90, 36, 114, 182, 146, 248, 113, 4, 0, 56,
-            39, 50, 64, 252, 29, 66, 90, 181, 133, 56, 26, 224, 99, 250, 83, 82,
-            149, 78, 118, 96, 32, 237, 87, 159, 217, 157, 194, 236, 62, 181,
-            104, 216, 16, 83, 160, 52, 232, 149, 2, 242, 123, 73, 120, 2, 111,
-            210, 152, 105, 159, 49, 174, 126, 170, 229, 41, 46, 115, 169, 19,
-            177, 85, 166, 253, 124, 88, 146, 107, 35, 4, 120, 108, 221, 250,
-            198, 132, 78, 186, 33, 22, 241, 234, 207, 220, 194, 90, 170, 208,
-            129, 19, 144, 194, 199, 71, 89, 206, 46, 249, 69, 83, 226, 179, 94,
-            216, 0, 103, 252, 72, 5, 108, 34, 197, 124, 3, 87, 31, 77, 100, 71,
-            102, 151, 1, 120, 223, 4, 218, 182, 162, 134, 220, 170, 0, 55, 151,
-            232, 25, 142, 230, 23, 63, 142, 211, 148, 157, 146, 112, 191, 108,
-            91, 228, 179, 104, 63, 145, 59, 157, 55, 25, 213, 221, 126, 234,
-            234, 188, 136, 214, 14, 90, 242, 176, 47, 248, 28, 16, 140, 39, 5,
-            172, 117, 109, 204, 106, 3, 239, 12, 35, 92, 202, 171, 75, 235, 204,
-            24, 9, 210, 177, 207, 119, 82, 144, 79, 186, 2, 101, 235, 69, 251,
-            145, 128, 52, 21, 208, 94, 99, 69, 215, 36, 107, 196, 108, 76, 206,
-            155, 120, 161, 233, 20, 182, 87, 228, 243, 187, 76, 163, 9, 3, 124,
-            246, 209, 14, 248, 145, 182, 222, 142, 243, 108, 111, 179, 126, 218,
-            58, 88, 246, 239, 152, 140, 20, 170, 8, 138, 211, 157, 209, 253,
-            185, 99, 177, 217, 139, 28, 119, 23, 103, 179, 104, 233, 101, 2,
-            166, 52, 117, 230, 207, 158, 121, 177, 217, 62, 244, 255, 171, 205,
-            146, 144, 203, 47, 77, 109, 0, 170, 54, 248, 121, 201, 187, 209, 9,
-            109, 3, 100, 204, 25, 149, 97, 178, 10, 198, 136, 83, 134, 98, 218,
-            219, 199, 70, 252, 190, 98, 75, 206, 102, 135, 173, 115, 175, 92,
-            186, 143, 60, 12, 145, 200, 16, 28, 209, 210, 39, 144, 141, 1, 200,
-            204, 217, 117, 52, 208, 139, 122, 235, 104, 152, 120, 101, 213, 161,
-            68, 156, 39, 207, 85, 162, 126, 163, 161, 212, 182, 49, 105, 189,
-            160, 249, 76, 220, 141, 237, 36, 6, 231, 91, 132, 241, 137, 194,
-            204, 61, 67, 33, 148, 238, 246, 166, 103, 73, 181, 46, 241, 49, 168,
-            202, 79, 165, 38, 209, 123, 204, 56, 42, 109, 90, 34, 12, 53, 180,
-            3, 2, 45, 52, 21, 170, 99, 150, 150, 180, 232, 253, 197, 241, 224,
-            233, 10, 129, 255, 2, 73, 168, 51, 121, 65, 251, 182, 239, 134, 7,
-            109, 44, 145, 140, 251, 221, 70, 49, 147, 73, 132, 200, 8, 242, 96,
-            192, 103, 82, 14
-          ],
+        const result = (await Promise.all(bulkRequest)).every(
+          (res) => res.data.valid === true
+        );
+
+        if (result === true) {
+          setIsLoading(false);
+          handleVerified();
+        } else {
+          setIsLoading(false);
+          handleNotVerified();
         }
-      );
-      if (res.data.valid === true) {
-        setIsLoading(false);
-
-        handleVerified();
-      }
-      else {
-        setIsLoading(false);
-        handleNotVerified();
       }
     } catch (e) {
-      setIsLoading(false);
       console.log(e);
     }
   };
@@ -387,14 +106,13 @@ export function UploadPanel() {
   const { isOpen, onToggle } = useDisclosure();
 
   const styles = useMultiStyleConfig("Button", { variant: "outline" });
-
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
   const [show, setShow] = React.useState(false);
   const handleToggle = () => setShow(!show);
 
   const [jsonFileContent, setJsonFileContent] =
-    React.useState<JsonFileContentType[]>([]);
+    React.useState<JsonFileContentType[]>();
 
   const uploadFile = async function (e: any) {
     setShowButton(true);
@@ -458,142 +176,167 @@ export function UploadPanel() {
           onChange={uploadFile}
         />
 
-        {jsonFileContent ? (
-          jsonFileContent.map((file) => {
-            // print file contents
-            // file.selectedRows.map((file1) => {
+        <div>
+          {jsonFileContent ? (
+            jsonFileContent.map((file) => {
+              // // print file contents
+              // file.selectedRows.map((file1) => {
+              //   // console.log(file1.selectedKey);
+              //   // console.log(file1.selectedValue);
+              //   const proofStr = file1.proof.toString().slice(0, 10) + "...";
 
-            //   // console.log(file1.selectedKey);
-            //   // console.log(file1.selectedValue);
-
-            //   const proofStr = file1.proof.toString().slice(0, 10) + "...";
-            //   console.log(`proof: ${proofStr}`);
-            // });
-            return (
-              <>
-                <Box mt="2" mb="2">
-                  <Card>
-                    <CardHeader>
-                      <Heading size="md">Proof</Heading>
-                    </CardHeader>
-                    <CardBody>
-                      <Stack divider={<StackDivider />} spacing="4">
-                        <Box>
-                          <Heading size="xs" textTransform="uppercase">
-                            Verifier Address
-                          </Heading>
-                          <Text pt="2" fontSize="sm">
-                            {address}
-                          </Text>
-                        </Box>
-                        <Box>
-                          <Heading size="xs" textTransform="uppercase">
-                            Address
-                          </Heading>
-                          <Text pt="2" fontSize="sm">
-                            {file.address}
-                          </Text>
-                        </Box>
-                        <Box>
-                          <Heading size="xs" textTransform="uppercase">
-                            Cert Name
-                          </Heading>
-                          <Text pt="2" fontSize="sm">
-                            {file.certName}
-                          </Text>
-                        </Box>
-                        <Box>
-                          <Heading size="xs" textTransform="uppercase">
-                            Cert Hash
-                          </Heading>
-                          <Text pt="2" fontSize="sm">
-                            {file.certHash}
-                          </Text>
-                        </Box>
-                        <Box>
-                          {file.selectedRows.map((file1) => {
-                            return (
-                              <>
-                                <Stack divider={<StackDivider />} spacing="4">
-                                  <Box>
-                                    <Heading
-                                      size="xs"
-                                      textTransform="uppercase"
-                                    >
-                                      Name
-                                    </Heading>
-                                    <Text pt="2" fontSize="sm">
-                                      {file1.selectedKey}
-                                    </Text>
-                                  </Box>
-                                  <Box>
-                                    <Heading
-                                      size="xs"
-                                      textTransform="uppercase"
-                                    >
-                                      Value
-                                    </Heading>
-                                    <Text pt="2" fontSize="sm">
-                                      {file1.selectedValue}
-                                    </Text>
-                                  </Box>
-                                  <Box>
-                                    <Heading
-                                      size="xs"
-                                      textTransform="uppercase"
-                                    >
-                                      Generated Proof
-                                    </Heading>
-                                    <Accordion allowToggle>
-                                      <AccordionItem borderColor={"white"}>
-                                        <AccordionButton>
-                                          <Box
-                                          // as="span"
-                                          // flex="1"
-                                          // textAlign="left"
-                                          >
-                                            <Heading
-                                              size="xs"
-                                              textTransform="uppercase"
-                                              className="border border-solid p-[10px] rounded-full bg-[#BDCDD6]"
+              //   console.log(`proof: ${proofStr}`);
+              // });
+              return (
+                <>
+                  <Box mt="2" mb="2">
+                    <Card>
+                      <CardHeader>
+                        <Heading className="font-black text-xl">
+                          Proof Details
+                        </Heading>
+                      </CardHeader>
+                      <CardBody>
+                        <Stack divider={<StackDivider />} spacing="4">
+                          <Box>
+                            <Heading
+                              size="xs"
+                              textTransform="uppercase"
+                              className="text-[#0F6292]"
+                            >
+                              Verifier Address
+                            </Heading>
+                            <Text pt="2" fontSize="sm">
+                              {address}
+                            </Text>
+                          </Box>
+                          <Box>
+                            <Heading
+                              size="xs"
+                              textTransform="uppercase"
+                              className="text-[#0F6292]"
+                            >
+                              Address
+                            </Heading>
+                            <Text pt="2" fontSize="sm">
+                              {file.address}
+                            </Text>
+                          </Box>
+                          <Box>
+                            <Heading
+                              size="xs"
+                              textTransform="uppercase"
+                              className="text-[#0F6292]"
+                            >
+                              Cert Name
+                            </Heading>
+                            <Text pt="2" fontSize="sm">
+                              {file.certName}
+                            </Text>
+                          </Box>
+                          <Box>
+                            <Heading
+                              size="xs"
+                              textTransform="uppercase"
+                              className="text-[#0F6292]"
+                            >
+                              Cert Hash
+                            </Heading>
+                            <Text pt="2" fontSize="sm">
+                              {file.certHash}
+                            </Text>
+                          </Box>
+                          <Box>
+                            {file.selectedRows.map((file1) => {
+                              return (
+                                <>
+                                  <Stack divider={<StackDivider />} spacing="4">
+                                    <Box>
+                                      <Heading
+                                        size="xs"
+                                        textTransform="uppercase"
+                                        className="text-[#0F6292]"
+                                      >
+                                        Name
+                                      </Heading>
+                                      <Text pt="2" fontSize="sm">
+                                        {file1.selectedKey}
+                                      </Text>
+                                    </Box>
+                                    <Box>
+                                      <Heading
+                                        size="xs"
+                                        textTransform="uppercase"
+                                        className="text-[#0F6292]"
+                                      >
+                                        Value
+                                      </Heading>
+                                      <Text pt="2" fontSize="sm">
+                                        {file1.selectedValue}
+                                      </Text>
+                                    </Box>
+                                    <Box>
+                                      <Heading
+                                        size="xs"
+                                        textTransform="uppercase"
+                                        className="text-[#0F6292]"
+                                      >
+                                        Generated Proof
+                                      </Heading>
+                                      <Accordion allowToggle>
+                                        <AccordionItem borderColor={"white"}>
+                                          <AccordionButton>
+                                            <Box
+                                            // as="span"
+                                            // flex="1"
+                                            // textAlign="left"
                                             >
-                                              Click to see Proof
-                                            </Heading>
-                                          </Box>
-                                          <AccordionIcon />
-                                        </AccordionButton>
-                                        <AccordionPanel pb={4}>
-                                          {file1.proof.toString()}
-                                        </AccordionPanel>
-                                      </AccordionItem>
-                                    </Accordion>
-                                  </Box>
-                                </Stack>
-                              </>
-                            );
-                          })}
-                        </Box>
-                      </Stack>
-                    </CardBody>
-                  </Card>
-                </Box>
-              </>
-            );
-          })
-        ) : (
-          <></>
-        )}
+                                              <Heading
+                                                size="xs"
+                                                textTransform="uppercase"
+                                                className="border border-solid p-[10px] rounded-full bg-[#BDCDD6]"
+                                              >
+                                                Click to see Proof
+                                              </Heading>
+                                            </Box>
+                                            <AccordionIcon />
+                                          </AccordionButton>
+                                          <AccordionPanel pb={4}>
+                                            {file1.proof.toString()}
+                                          </AccordionPanel>
+                                        </AccordionItem>
+                                      </Accordion>
+                                    </Box>
+                                  </Stack>
+                                </>
+                              );
+                            })}
+                          </Box>
+                        </Stack>
+                      </CardBody>
+                    </Card>
+                  </Box>
+                </>
+              );
+            })
+          ) : (
+            <></>
+          )}
+        </div>
+
       </div>
-      
+
       {showButton && (
         <Button
           isLoading={isLoading}
           onClick={handleVerifyButton}
-          className=" w-[150px] h-[50px] my-[20px]"
+          className="border boder-solid rounded-full 
+      w-[150px] h-[50px] font-bold bg-lime-200 my-[20px]"
         >
           Verify Proof
         </Button>
       )}
+
     </div>
   );
 }
